% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPFF.R
\docType{data}
\name{GSPFF}
\alias{GSPFF}
\title{Aboveground traits from the global spectrum of plant form and function (complete data)}
\format{
## `GSPFF`
A data frame with 2,630 rows and 6 columns:
\describe{
  \item{la}{leaf area}
  \item{ln}{leaf nitrogen content}
  \item{ph}{plant height}
  \item{sla}{specific leaf area}
  \item{ssd}{specific stem density}
  \item{sm}{seed mass}
  ...
}
}
\source{
<https://doi.org/10.6084/m9.figshare.13140146>
}
\usage{
GSPFF
}
\description{
Data on six aboveground traits for 2,630 species with complete trait information. Data was processed from the TRY database (https://www.try-db.org/TryWeb/Home.php) and used in the paper "Fine-root traits in the global spectrum of plant form and function (Carmona et al. 2021, Nature). The data is available in https://doi.org/10.6084/m9.figshare.13140146. All traits are log10 transformed and scaled.
}
\keyword{datasets}
