% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPFF_missing.R
\docType{data}
\name{GSPFF_missing}
\alias{GSPFF_missing}
\title{Aboveground traits from the global spectrum of plant form and function (incomplete data)}
\format{
## `GSPFF_missing`
A data frame with 10,746 rows and 6 columns:
\describe{
  \item{la}{leaf area}
  \item{ln}{leaf nitrogen content}
  \item{ph}{plant height}
  \item{sla}{specific leaf area}
  \item{ssd}{specific stem density}
  \item{sm}{seed mass}
  ...
}
}
\source{
<https://doi.org/10.6084/m9.figshare.13140146>
}
\usage{
GSPFF_missing
}
\description{
Data on six aboveground traits for 10,746 species with incomplete trait information. Data was processed from the TRY database (https://www.try-db.org/TryWeb/Home.php) and used in the paper "Fine-root traits in the global spectrum of plant form and function (Carmona et al. 2021, Nature). The data is available in https://doi.org/10.6084/m9.figshare.13140146. Only species with information for at least three traits are included. All traits are log10 transformed and scaled.
}
\keyword{datasets}
