% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPFF_missing_tax.R
\docType{data}
\name{GSPFF_missing_tax}
\alias{GSPFF_missing_tax}
\title{Taxonomic information for plants from the global spectrum of plant form and function (incomplete data)}
\format{
## `GSPFF_missing_tax`
A data frame with 10,746 rows and 3 columns:
\describe{
  \item{genus}{genus to which the species belongs}
  \item{family}{family to which the species belongs}
  \item{order}{order to which the species belongs}
  ...
}
}
\source{
<https://doi.org/10.6084/m9.figshare.13140146>
}
\usage{
GSPFF_missing_tax
}
\description{
Taxonomic data for 10,746 species with incomplete trait information (species with at least three traits).
}
\keyword{datasets}
