% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\docType{data}
\name{phylo}
\alias{phylo}
\title{Phylogeny for species from the global spectrum of plant form and function (incomplete data)}
\format{
## `phylo`
An object of class \code{"phylo"}
}
\usage{
phylo
}
\description{
Phylogenetic tree including information for 10,746 species with incomplete trait information (species with at least three traits), contained in \code{GSPFF_missing}.
}
\keyword{datasets}
